auth_service/src/requirements.txt
aenum==3.1.15 ; python_version >= "3.10" and python_version < "4.0"
alembic==1.12.0 ; python_version >= "3.10" and python_version < "4.0"
annotated-types==0.6.0 ; python_version >= "3.10" and python_version < "4.0"
anyio==4.2.0 ; python_version >= "3.10" and python_version < "4.0"
asgi-correlation-id==4.3.0 ; python_version >= "3.10" and python_version < "4.0"
asgiref==3.7.2 ; python_version >= "3.10" and python_version < "4.0"
async-timeout==4.0.3 ; python_version >= "3.10" and python_version < "4.0"
asyncpg==0.28.0 ; python_version >= "3.10" and python_version < "4.0"
certifi==2023.11.17 ; python_version >= "3.10" and python_version < "4.0"
cffi==1.16.0 ; python_version >= "3.10" and python_version < "4.0"
click==7.1.2 ; python_version >= "3.10" and python_version < "4.0"
cryptography==41.0.7 ; python_version >= "3.10" and python_version < "4.0"
deprecated==1.2.14 ; python_version >= "3.10" and python_version < "4.0"
dnspython==2.4.2 ; python_version >= "3.10" and python_version < "4.0"
email-validator==2.1.0.post1 ; python_version >= "3.10" and python_version < "4.0"
exceptiongroup==1.2.0 ; python_version >= "3.10" and python_version < "3.11"
fastapi-limiter==0.1.5 ; python_version >= "3.10" and python_version < "4.0"
fastapi==0.108.0 ; python_version >= "3.10" and python_version < "4.0"
googleapis-common-protos==1.56.2 ; python_version >= "3.10" and python_version < "4.0"
greenlet==3.0.3 ; python_version >= "3.10" and python_version < "4.0" and (platform_machine == "aarch64" or platform_machine == "ppc64le" or platform_machine == "x86_64" or platform_machine == "amd64" or platform_machine == "AMD64" or platform_machine == "win32" or platform_machine == "WIN32")
grpcio==1.60.0 ; python_version >= "3.10" and python_version < "4.0"
gunicorn==21.2.0 ; python_version >= "3.10" and python_version < "4.0"
h11==0.14.0 ; python_version >= "3.10" and python_version < "4.0"
h2==4.1.0 ; python_version >= "3.10" and python_version < "4.0"
hpack==4.0.0 ; python_version >= "3.10" and python_version < "4.0"
httpcore==1.0.2 ; python_version >= "3.10" and python_version < "4.0"
httptools==0.6.0 ; python_version >= "3.10" and python_version < "4.0"
httpx[http2]==0.25.1 ; python_version >= "3.10" and python_version < "4.0"
hyperframe==6.0.1 ; python_version >= "3.10" and python_version < "4.0"
idna==3.6 ; python_version >= "3.10" and python_version < "4.0"
importlib-metadata==6.0.1 ; python_version >= "3.10" and python_version < "4.0"
mako==1.3.0 ; python_version >= "3.10" and python_version < "4.0"
markupsafe==2.1.3 ; python_version >= "3.10" and python_version < "4.0"
opentelemetry-api==1.17.0 ; python_version >= "3.10" and python_version < "4.0"
opentelemetry-exporter-jaeger-proto-grpc==1.17.0 ; python_version >= "3.10" and python_version < "4.0"
opentelemetry-exporter-jaeger-thrift==1.17.0 ; python_version >= "3.10" and python_version < "4.0"
opentelemetry-exporter-jaeger==1.17.0 ; python_version >= "3.10" and python_version < "4.0"
opentelemetry-instrumentation-asgi==0.38b0 ; python_version >= "3.10" and python_version < "4.0"
opentelemetry-instrumentation-fastapi==0.38b0 ; python_version >= "3.10" and python_version < "4.0"
opentelemetry-instrumentation==0.38b0 ; python_version >= "3.10" and python_version < "4.0"
opentelemetry-sdk==1.17.0 ; python_version >= "3.10" and python_version < "4.0"
opentelemetry-semantic-conventions==0.38b0 ; python_version >= "3.10" and python_version < "4.0"
opentelemetry-util-http==0.38b0 ; python_version >= "3.10" and python_version < "4.0"
orjson==3.8.7 ; python_version >= "3.10" and python_version < "4.0"
packaging==23.2 ; python_version >= "3.10" and python_version < "4.0"
passlib==1.7.4 ; python_version >= "3.10" and python_version < "4.0"
protobuf==3.20.3 ; python_version >= "3.10" and python_version < "4.0"
psycopg2-binary==2.9.9 ; python_version >= "3.10" and python_version < "4.0"
pycparser==2.21 ; python_version >= "3.10" and python_version < "4.0"
pydantic-core==2.14.6 ; python_version >= "3.10" and python_version < "4.0"
pydantic-settings==2.1.0 ; python_version >= "3.10" and python_version < "4.0"
pydantic==2.5.3 ; python_version >= "3.10" and python_version < "4.0"
pydantic[email]==2.5.3 ; python_version >= "3.10" and python_version < "4.0"
pyjwt[crypto]==2.8.0 ; python_version >= "3.10" and python_version < "4.0"
python-dotenv==1.0.0 ; python_version >= "3.10" and python_version < "4.0"
redis==4.4.4 ; python_version >= "3.10" and python_version < "4.0"
setuptools==69.0.3 ; python_version >= "3.10" and python_version < "4.0"
six==1.16.0 ; python_version >= "3.10" and python_version < "4.0"
sniffio==1.3.0 ; python_version >= "3.10" and python_version < "4.0"
sqlalchemy==2.0.25 ; python_version >= "3.10" and python_version < "4.0"
sqlmodel==0.0.14 ; python_version >= "3.10" and python_version < "4.0"
starlette==0.32.0.post1 ; python_version >= "3.10" and python_version < "4.0"
structlog==24.1.0 ; python_version >= "3.10" and python_version < "4.0"
thrift==0.16.0 ; python_version >= "3.10" and python_version < "4.0"
typer==0.9.0 ; python_version >= "3.10" and python_version < "4.0"
typing-extensions==4.9.0 ; python_version >= "3.10" and python_version < "4.0"
uvicorn==0.12.2 ; python_version >= "3.10" and python_version < "4.0"
uvloop==0.17.0 ; python_version >= "3.10" and python_version < "4.0"
wrapt==1.16.0 ; python_version >= "3.10" and python_version < "4.0"
zipp==3.17.0 ; python_version >= "3.10" and python_version < "4.0"
